Fix ZLib-Intel Build Warnings

+++ src/Native/Windows/clrcompression/zlib-intel/inflate.c
@@ -505,7 +505,7 @@ unsigned copy;
 /* Restore state from registers in inflate() */
 #define RESTORE() \
     do { \
-        state->wnext = put - (state->window + state->wsize);\
+        state->wnext = (unsigned)(put - (state->window + state->wsize));\
         strm->avail_out = left; \
         strm->next_in = next; \
         strm->avail_in = have; \
@@ -938,7 +938,7 @@ int flush;
             copy = state->length;
             if (copy) {
                 unsigned char *end = state->window + (state->wsize * 4);
-                unsigned long diff = end - put;
+                unsigned diff = (unsigned)(end - put);
 
                 if (copy > have) copy = have;
                 if (copy > diff) {
@@ -946,7 +946,7 @@ int flush;
                         RESTORE();
                         window_output_flush(strm);
                         LOAD();
-                        diff = end - put;
+                        diff = (unsigned)(end - put);
                     }
                     if (copy > diff) copy = diff;
                 }
@@ -1189,7 +1189,7 @@ int flush;
             state->mode = MATCH;
             case MATCH: {
                 unsigned char *end = state->window + (state->wsize * 4);
-                unsigned long buf_left = end - put;
+                unsigned buf_left = (unsigned)(end - put);
                 copy = state->length;
 
 		RESTORE();
@@ -1198,7 +1198,7 @@ int flush;
                         /* relies on RESTORE() above with no changes to those vars */
                         window_output_flush(strm);
                         LOAD();
-                        buf_left = end - put;
+                        buf_left = (unsigned)(end - put);
                     }
 		    if (copy > buf_left) copy = buf_left;
                 }
